refactor: 패키지 이름 변경 및 위치 이동 #8
[domain 패키지 -> entity 패키지] 이름 변경,
mapper 패키지 위치 이동
